requirements.txt
Django==3.2.25
django-cors-headers==4.5.0
django-filter==23.5
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.7.0
gunicorn==23.0.0
pymongo==3.12.1